slt/sltu with limb style circuit
