+++ ceno_zkvm/src/gadgets/mod.rs
@@ -2,6 +2,7 @@ mod div;
 mod is_lt;
 mod signed;
 mod signed_ext;
+mod signed_limbs;
 
 pub use div::DivConfig;
 pub use gkr_iop::gadgets::{
@@ -10,3 +11,4 @@ pub use gkr_iop::gadgets::{
 pub use is_lt::{AssertSignedLtConfig, SignedLtConfig};
 pub use signed::Signed;
 pub use signed_ext::SignedExtendConfig;
+pub use signed_limbs::{UIntLimbsLT, UIntLimbsLTConfig};
